requirements.txt
fastapi
uvicorn
starkbank
python-dotenv
apscheduler
faker